Upgrades Galore (Nuget, .NET .NET 4.7.2, etc.; w/o debugging)
* Remove obsolete project: It looks like we no longer use nunitforms.
* Remove unused or obsolete files
** Remove unused nmock source code
** Remove .idea files
** Remove obsolete CommonCOMDlgs and Widgets
    CommonCOMDlgs doesn't seem to be used anymore - at least I
	couldn't find any references to it's classes anywhere in the Src
	directory (other than in CommonCOMDlgs itself). Same is true for
	Widgets.
** Remove obsolete batch files.
** Remove unused properties and items
** Remove ParaTExt 7
* Consume Nuget packages
* Update output directory in Code launch.json
* Remove warnings
* Set net472 output path in environ to find and load libgeckofix.so
* Update the output directory in _init.mak and environ to net472
* Fix missing dispose of FwResourceHelperImpl when running tests
* Update packages to avoid icu dll copy issue
* Copy the .idh files from liblcm to Kernel.
    This will let the Kernel project build on Windows
* Update TLB related files for Linux
* Add C++ projects to solution
* Update unit++.vcxproj to toolset v142
* Move common properties to Directory.Build.props
* Add MGAStrings Designer back into LanguageExplorer
* Untangle circular dependency
    When we moved some of the tasks from BuildInclude.targets into
	ViewsInterfaces we created a circular dependency.
    The Views target generates ViewsTbl.idl.
    ViewsInterfaces needs ViewsTbl.idl to run IDLImport.
    Views depends on ViewsInterfaces.

    This change calls the part of ViewsInterfaces that we need to build
    FwKernel and Views separatedly before building ViewsInterfaces.
* Adjust NUnitReport to new output location
    This fixes sillsdev/libpalaso#1038.
* Move FwBuildTasks to separate project
    FwBuildTasks is now a separate project [1] that creates it's own
    nuget package, so we don't have to commit FwBuildTasks.dll anymore
    nor do we need the source here.

[1] https://github.com/sillsdev/SIL.FwBuildTasks

WIP: Get the build working - Works on Windows and Linux!

 * To Get tests working
 - Copy LcmTemplateFiles
 - Add AutomaticBindingRedirect to test projects
 * Fix RegFree target for people who have spaces in their usernames
 * Copy MetaDataCache test files to output folder for use in unit tests
 * Adjust ParatextImport tests to match NUnit3 behavior

Build C++ projects on Linux from vcxproj without VS

- Don't import VCTargetsPath files if don't exist.
- Use empty Restore task.
- For Build task, run the Makefile.
- unit++.vcxproj can use autoreconf like the root Makefile is doing.
- {Kernel,views}/Makefile: Changing directory and then relying on
  the OUT_DIR relative path being correct isn't working. Perhaps
  because the OUT_DIR is now one level deeper. But we already
  `cd $(OUT_DIR)` so just omit `$(OUT_DIR)/` from the path of the
  file being generated.
- _init.mak.lnx:
  - When Makefiles run generate-map.sh, it needs to be found in PATH.
    But there is trouble since LIBCOM_DIR is a relative path (and
    OUT_DIR is different now with `net462`). Make LIBCOM_DIR be
    absolute.
  - Provide the path to FW ICU so we don't need to have first gotten
    it from `environ`.
  - Provide libcom and FW ICU in LD_LIBRARY_PATH, so we don't need to
    have first gotten it from `environ`, to run generate-map.sh.

Fix lame settings on Linux

The previous settings output a warning on Linux because -B requires
an argument, and it spit out a lot of unnecessary output. This
change removes the warning by omitting the -B parameter, and
suppresses the output.

Fix some hanging tests

* Fix possible teardown crashes in LcmLexiconTests
* Re-Enable the InstallValidatorTests
* Use MessageBoxUtils in GlossListTreeView to avoid message boxes
  popping up in unit tests
* Add Paratext9 support to ProjectUnpacker
* Also fix unit test file load and hanging issue In SCTextEnumTests
* Skip initting ICU for tests (it causes them to hang)

+++ Src/ParatextImportTests/DiffTestHelper.cs
@@ -163,8 +163,8 @@ namespace ParatextImport
 		{
 			Difference subDiff = rootDiff.SubDiffsForORCs[iSubDiff];
 			// verify the basics
-			Assert.AreEqual(0, subDiff.RefStart);
-			Assert.AreEqual(0, subDiff.RefEnd);
+			Assert.AreEqual(0, (int)subDiff.RefStart);
+			Assert.AreEqual(0, (int)subDiff.RefEnd);
 			Assert.AreEqual(DifferenceType.NoDifference, subDiff.DiffType);
 
 			// the Current para stuff
@@ -205,8 +205,8 @@ namespace ParatextImport
 		{
 			Difference subDiff = rootDiff.SubDiffsForORCs[iSubDiff];
 			// verify the basics
-			Assert.AreEqual(0, subDiff.RefStart);
-			Assert.AreEqual(0, subDiff.RefEnd);
+			Assert.AreEqual(0, (int)subDiff.RefStart);
+			Assert.AreEqual(0, (int)subDiff.RefEnd);
 			Assert.AreEqual(DifferenceType.NoDifference, subDiff.DiffType);
 
 			// the Current para stuff
@@ -705,8 +705,8 @@ namespace ParatextImport
 			string verseText, int iVerseStart, bool fIsChapter, bool fIsHeading, int iSection)
 		{
 			Assert.AreEqual(para, scrVerse.Para);
-			Assert.AreEqual(startRef, scrVerse.StartRef);
-			Assert.AreEqual(endRef, scrVerse.EndRef);
+			Assert.AreEqual(startRef, (int)scrVerse.StartRef);
+			Assert.AreEqual(endRef, (int)scrVerse.EndRef);
 			Assert.AreEqual(verseText, scrVerse.Text.Text);
 			Assert.AreEqual(iVerseStart, scrVerse.VerseStartIndex);
 			Assert.AreEqual(fIsChapter, scrVerse.ChapterNumberRun);
